Added marshaling of weights for the Logical map. Added comparison operators for easier testing. Fixed marshaling bugs in the Occupancy and Logical maps. Added tests around bugs and new features. Discovered and fixed a bug in A*.

+++ srs_chuck/src/srslib_framework/include/srslib_framework/planning/pathplanning/grid2d/Grid2dSolutionUtils.hpp
@@ -17,7 +17,7 @@ namespace srs {
 
 struct Grid2dSolutionUtils
 {
-    static double getTotalCost(Solution<Grid2dSolutionItem>& solution);
+    static int getTotalCost(Solution<Grid2dSolutionItem>& solution);
     static vector<Solution<Grid2dSolutionItem>*> splitSolution(Solution<Grid2dSolutionItem>& solution);
 };
